Address #1648 by breaking replacePathVariables into TemplatedPathHelper.js
